requirements_dev.txt
pyapriltags
numpy
opencv-contrib-python
flask
robotpy-wpimath
pyntcore

mcap
protobuf3
mcap-protobuf-support
foxglove-schemas-protobuf
foxglove-websocket
typing
asyncio
scipy
